Make configuration type-safe
This is a rather huge commit which changes arara's configuration
system to the type-safe `konf` model. In the course of this major
refactoring nearly nothing has been touched, especially we have many
dangling potential duplicates which have to be addressed later on.

While introducing this type-safe configuration, the `max-loops` key's
type has been reduced from long to integer. I do not suspect that this
will break any workflow.

Furthermore, the application now uses the experimental `kotlin.time`
API even in its configuration. We have used it before (in the main
method), so we would have to refactor anyway if it shouldn't stabilize.