Add option to change working directory
This commit introduces working directory support. It defaults to the
current directory and may be changed arbitrarily.

Please note: All file arguments passed to arara will be resolved
against the working directory. This is justified by the TeX engines
being called within the working directory and thus searching the same
files. Please note that it may require different path handling for the
user on the command-line calling arara from another directory.

This fixes #13 and the build failures on JDK 11.