Refactor and cleanup
This commit tries to improve type safety by using file methods for
file and path building instead of mapping to string and back. This is
not completed yet but we are on the right track removing string based
path manipulation.

Furthermore, some uses of `List` have been replaced by `Set` to
emphasize that they will not contain duplicates.