Remove `getFullBasename` and `addQuotes`
`getFullBasename` has provided very inconsistent support for getting a
full path with the basename of the file appended. As the user can access
the Java API, it will be easy to implement a working method for the use
case.

`addQuotes` only double quoted the parameter. That is not necessary, it
is shorter for the user to write double quotes.