Move towards abstract LocalConfiguration
This commit moves some relevant properties into the API so that
arara's rule paths and file type configuration may be
configured. That includes moving some defaults to the core
implementation.

The most important step is the reduction of code distribution where a
function is called only once in the project. LocalConfiguration and
Configuration are a prime example for this and have been refactored.