Remove RuleErrorHeader and centrally catch exception
This commit replaces the RuleErrorHeader by another type of exception which may
only be triggered while executing Interpreter tasks. By catching at one place,
arara does not need to keep the state of the current rule around (which would be
a problem for parallel execution anyway) and does the formatting on the fly.

In the course of this change, `AraraException` has been opened and
`HaltExpectedException` takes advantage of this and is now a subtype of the
former.